nebula-cli/cli/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nebula-cli"
version = "0.1.0"
description = "CLI for the Nebula Cloud System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]",
    "requests",
    "python-dotenv",
    "rich"
]

[project.scripts]
nebula = "src.main:app"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]